+++ Changelog.md
@@ -12,6 +12,7 @@
 * Redone parsing Bookmarks, since it part of `DocxParagraph` elements and order is important
 * `BookmarkStart` and `BookmarkEnd` count as nonempty_run
 * Add parsing of `CommentRangeStart` and `CommentRangeEnd`
+* Add parsing of `Series#values`
 
 ### Refactor
 * Store `sdt` as `DocxParagraph#character_styles_array` element